+++ src/predict.cxx
@@ -6,10 +6,10 @@
 namespace performance_model{
 
 double predict(const double* configuration, const model* interpolator, const model* extrapolator){
-  int nparam = extrapolator->get_num_inputs();
+  size_t nparam = extrapolator->get_num_inputs();
   assert(nparam == extrapolator->get_num_inputs());
   bool use_interpolation_model = true;
-  for (int i=0; i<nparam; i++){
+  for (size_t i=0; i<nparam; i++){
     if (interpolator->get_max_observed_parameter_value(i) < configuration[i] || interpolator->get_min_observed_parameter_value(i) > configuration[i]){
       use_interpolation_model = false;
       break;
